sample_python_script/requirements.txt
pandas==0.19.2
pyserial==3.2.1